MessageDef: enforce format for PrettyPrintIM and fix usage of subscriptionId after size change

+++ examples/tv-casting-app/tv-casting-common/commands/clusters/SubscriptionsCommands.h
@@ -26,7 +26,7 @@ class ShutdownSubscription : public CHIPCommand
 public:
     ShutdownSubscription(CredentialIssuerCommands * credsIssuerConfig) : CHIPCommand("shutdown-subscription", credsIssuerConfig)
     {
-        AddArgument("subscription-id", 0, UINT64_MAX, &mSubscriptionId);
+        AddArgument("subscription-id", 0, UINT32_MAX, &mSubscriptionId);
     }
 
     /////////// CHIPCommand Interface /////////
